requirements.txt
tweepy==4.15.0
hyperliquid_python_sdk==0.12.0
python-dotenv==1.1.0
emp-agents==0.3.3
httpx==0.27.2
